requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0
httpx>=0.27.0
sqlalchemy>=2.0.25
sqlalchemy-bigquery>=1.9.0
pyarrow>=14.0.0
google-cloud-bigquery>=3.14.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
redis>=5.0.0
python-dotenv>=1.0.0
